도커 컨테이너 및 이미지 관련 스크립트 수정 및 개선

+++ publish.sh
@@ -1,6 +1,19 @@
 #!/bin/bash
-docker cp silgerae:/opt/myproject/db/silgerae.db /volume1/homes/brokim/silgerae/db
-docker rm -f silgerae
-docker rmi silgerae2
+# check if the container is running
+if [ "$(docker ps -q -f name=silgerae)" ]; then
+    echo "=> Copying the database file to the host"
+    docker cp silgerae:/opt/myproject/db/silgerae.db /volume1/homes/brokim/silgerae/db
+    echo "=> Removing the container"
+    docker rm -f silgerae
+fi
+# check if the image exists
+if [ "$(docker images -q silgerae)" ]; then
+    echo "=> Removing the image"
+    docker rmi silgerae2
+fi
+git pull
+echo "=> Building the image"
 docker build -t silgerae --name Dockerfile_silgerae .
+echo "=> Running the container"
 docker run -d -p 8000:8000 --name silgerae2 silgerae
+echo "=> Done"
